+++ lib/src/model/vimeo_video_config.dart
@@ -19,13 +19,31 @@ class VimeoRequest {
 class VimeoFiles {
   VimeoFiles({this.progressive});
 
-  factory VimeoFiles.fromJson(Map<String, dynamic> json) => VimeoFiles(
-        progressive: List<VimeoProgressive>.from(
-          json["progressive"].map(
-            (x) => VimeoProgressive.fromJson(x),
+  factory VimeoFiles.fromJson(Map<String, dynamic> json) {
+    final jsonData = json['progressive'] as List<dynamic>;
+
+    final progressiveUrls = List.generate(
+      jsonData.length,
+      (index) => VimeoProgressive.fromJson(jsonData[index]),
+    );
+
+    if (progressiveUrls.isEmpty) {
+      final jsonRes = json['hls']['cdns'];
+
+      for (final element in (jsonRes as Map).entries.toList()) {
+        progressiveUrls.add(
+          VimeoProgressive(
+            url: element.value['url'] as String,
           ),
-        ),
-      );
+        );
+        break;
+      }
+    }
+
+    return VimeoFiles(
+      progressive: progressiveUrls,
+    );
+  }
 
   List<VimeoProgressive?>? progressive;
 }